+++ Demo/Demo/GiphyAPI.swift
@@ -18,13 +18,13 @@ extension Giphy: TargetType {
     public var method: Moya.Method {
         switch self {
         case .upload:
-            return .POST
+            return .post
         }
     }
     public var parameters: [String: Any]? {
         switch self {
         case .upload:
-            return ["api_key": "dc6zaTOxFJmzC" as AnyObject, "username": "Moya" as AnyObject]
+            return ["api_key": "dc6zaTOxFJmzC", "username": "Moya"]
         }
     }
     public var task: Task {