+++ Demo/Tests/EndpointSpec.swift
@@ -11,16 +11,16 @@ class EndpointSpec: QuickSpec {
                 let target: GitHub = .zen
                 let parameters = ["Nemesis": "Harvey"]
                 let headerFields = ["Title": "Dominar"]
-                endpoint = Endpoint<GitHub>(URL: url(target), sampleResponseClosure: {.networkResponse(200, target.sampleData)}, method: Moya.Method.GET, parameters: parameters, parameterEncoding: JSONEncoding(), httpHeaderFields: headerFields)
+                endpoint = Endpoint<GitHub>(URL: url(target), sampleResponseClosure: {.networkResponse(200, target.sampleData)}, method: Moya.Method.get, parameters: parameters, parameterEncoding: JSONEncoding.default, httpHeaderFields: headerFields)
             }
             
             it("returns a new endpoint for endpointByAddingParameters") {
                 let message = "I hate it when villains quote Shakespeare."
-                let newEndpoint = endpoint.endpointByAddingParameters(["message": message])
+                let newEndpoint = endpoint.adding(newParameters: ["message": message])
                 let newEndpointMessageObject: Any? = newEndpoint.parameters?["message"]
                 let newEndpointMessage = newEndpointMessageObject as? String
-                let encodedRequest = try? endpoint.parameterEncoding.encode(newEndpoint.urlRequest, with: newEndpoint.parameters)
-                let newEncodedRequest = try? newEndpoint.parameterEncoding.encode(newEndpoint.urlRequest, with: newEndpoint.parameters)
+                let encodedRequest = try? endpoint.parameterEncoding.encode(newEndpoint.urlRequest!, with: newEndpoint.parameters)
+                let newEncodedRequest = try? newEndpoint.parameterEncoding.encode(newEndpoint.urlRequest!, with: newEndpoint.parameters)
                 
                 
                 // Make sure our closure updated the sample response, as proof that it can modify the Endpoint
@@ -35,10 +35,10 @@ class EndpointSpec: QuickSpec {
             
             it("returns a new endpoint for endpointByAddingHTTPHeaderFields") {
                 let agent = "Zalbinian"
-                let newEndpoint = endpoint.endpointByAddingHTTPHeaderFields(["User-Agent": agent])
+                let newEndpoint = endpoint.adding(newHttpHeaderFields: ["User-Agent": agent])
                 let newEndpointAgent = newEndpoint.httpHeaderFields?["User-Agent"]
-                let encodedRequest = try? endpoint.parameterEncoding.encode(newEndpoint.urlRequest, with: newEndpoint.parameters)
-                let newEncodedRequest = try? newEndpoint.parameterEncoding.encode(newEndpoint.urlRequest, with: newEndpoint.parameters)
+                let encodedRequest = try? endpoint.parameterEncoding.encode(newEndpoint.urlRequest!, with: newEndpoint.parameters)
+                let newEncodedRequest = try? newEndpoint.parameterEncoding.encode(newEndpoint.urlRequest!, with: newEndpoint.parameters)
                 
                 // Make sure our closure updated the sample response, as proof that it can modify the Endpoint
                 expect(newEndpointAgent).to(equal(agent))
@@ -51,10 +51,10 @@ class EndpointSpec: QuickSpec {
             }
 
             it ("returns a new endpoint for endpointByAddingParameterEncoding") {
-                let parameterEncoding = JSONEncoding()
-                let newEndpoint = endpoint.endpointByAddingParameterEncoding(parameterEncoding)
-                let encodedRequest = try? parameterEncoding.encode(newEndpoint.urlRequest, with: newEndpoint.parameters)
-                let newEncodedRequest = try? newEndpoint.parameterEncoding.encode(newEndpoint.urlRequest, with: newEndpoint.parameters)
+                let parameterEncoding = JSONEncoding.default
+                let newEndpoint = endpoint.adding(newParameterEncoding: parameterEncoding)
+                let encodedRequest = try? parameterEncoding.encode(newEndpoint.urlRequest!, with: newEndpoint.parameters)
+                let newEncodedRequest = try? newEndpoint.parameterEncoding.encode(newEndpoint.urlRequest!, with: newEndpoint.parameters)
 
                 // Make sure we updated the parameter encoding
                 expect(newEncodedRequest).to(equal(encodedRequest))
@@ -67,16 +67,16 @@ class EndpointSpec: QuickSpec {
             }
             
             it ("returns a new endpoint for endpointByAdding with all parameters") {
-                let parameterEncoding = URLEncoding()
+                let parameterEncoding = URLEncoding.default
                 let agent = "Zalbinian"
                 let message = "I hate it when villains quote Shakespeare."
-                let newEndpoint = endpoint.endpointByAdding(
+                let newEndpoint = endpoint.adding(
                     parameters: ["message": message],
                     httpHeaderFields: ["User-Agent": agent],
                     parameterEncoding: parameterEncoding
                 )
-                let encodedRequest = try? parameterEncoding.encode(newEndpoint.urlRequest, with: newEndpoint.parameters)
-                let newEncodedRequest = try? newEndpoint.parameterEncoding.encode(newEndpoint.urlRequest, with: newEndpoint.parameters)
+                let encodedRequest = try? parameterEncoding.encode(newEndpoint.urlRequest!, with: newEndpoint.parameters)
+                let newEncodedRequest = try? newEndpoint.parameterEncoding.encode(newEndpoint.urlRequest!, with: newEndpoint.parameters)
                 
                 
                 let newEndpointAgent = newEndpoint.httpHeaderFields?["User-Agent"]
@@ -96,6 +96,25 @@ class EndpointSpec: QuickSpec {
                 let title = titleObject as? String
                 expect(title).to(equal("Dominar"))
             }
+
+            it("returns a nil urlRequest for an invalid URL") {
+                let badEndpoint = Endpoint<Empty>(URL: "some invalid URL", sampleResponseClosure: { .networkResponse(200, Data()) })
+
+                expect(badEndpoint.urlRequest).to( beNil() )
+            }
         }
     }
+}
+
+enum Empty {
+}
+
+extension Empty: TargetType {
+    // None of these matter since the Empty has no cases and can't be instantiated.
+    var baseURL: URL { return URL(string: "http://example.com")! }
+    var path: String { return "" }
+    var method: Moya.Method { return .get }
+    var parameters: [String: Any]? { return nil }
+    var task: Task { return .request }
+    var sampleData: Data { return Data() }
 }