+++ Demo/Tests/MethodSpec.swift
@@ -6,15 +6,15 @@ class MethodSpec: QuickSpec {
     override func spec() {
         describe("supportsMultipart") {
             let expectations: [(Moya.Method, Bool)] = [
-                (.GET, false),
-                (.POST, true),
-                (.PUT, true),
-                (.DELETE, false),
-                (.OPTIONS, false),
-                (.HEAD, false),
-                (.PATCH, true),
-                (.TRACE, false),
-                (.CONNECT, true),
+                (.get, false),
+                (.post, true),
+                (.put, true),
+                (.delete, false),
+                (.options, false),
+                (.head, false),
+                (.patch, true),
+                (.trace, false),
+                (.connect, true),
             ]
             for (method, expected) in expectations {
                 it("\(method) should \(expected ? "" : "not") support multipart") {