+++ Demo/Tests/TestHelpers.swift
@@ -24,7 +24,7 @@ extension GitHub: TargetType {
     }
     
     var method: Moya.Method {
-        return .GET
+        return .get
     }
     
     var parameters: [String: Any]? {
@@ -66,7 +66,7 @@ enum HTTPBin: TargetType {
     }
 
     var method: Moya.Method {
-        return .GET
+        return .get
     }
     
     var parameters: [String: Any]? {
@@ -86,4 +86,53 @@ enum HTTPBin: TargetType {
             return "{\"authenticated\": true, \"user\": \"user\"}".data(using: String.Encoding.utf8)!
         }
     }
+}
+
+public enum GitHubUserContent {
+    case downloadMoyaWebContent(String)
+}
+
+extension GitHubUserContent: TargetType {
+    public var baseURL: URL { return URL(string: "https://raw.githubusercontent.com")! }
+    public var path: String {
+        switch self {
+        case .downloadMoyaWebContent(let contentPath):
+            return "/Moya/Moya/master/web/\(contentPath)"
+        }
+    }
+    public var method: Moya.Method {
+        switch self {
+        case .downloadMoyaWebContent:
+            return .get
+        }
+    }
+    public var parameters: [String: Any]? {
+        switch self {
+        case .downloadMoyaWebContent:
+            return nil
+        }
+    }
+    public var task: Task {
+        switch self {
+        case .downloadMoyaWebContent:
+            return .download(.request(DefaultDownloadDestination))
+        }
+    }
+    public var sampleData: Data {
+        switch self {
+        case .downloadMoyaWebContent:
+            return Data(count: 4000)
+        }
+    }
+   
+}
+
+private let DefaultDownloadDestination: DownloadDestination = { temporaryURL, response in
+    let directoryURLs = FileManager.default.urls(for: .cachesDirectory, in: .userDomainMask)
+
+    if !directoryURLs.isEmpty {
+        return (directoryURLs.first!.appendingPathComponent(response.suggestedFilename!), [])
+    }
+    
+    return (temporaryURL, [])
 }