+++ Source/Moya+Alamofire.swift
@@ -6,7 +6,6 @@ internal typealias Request = Alamofire.Request
 internal typealias DownloadRequest = Alamofire.DownloadRequest
 internal typealias UploadRequest = Alamofire.UploadRequest
 internal typealias DataRequest = Alamofire.DataRequest
-internal typealias StreamRequest = Alamofire.StreamRequest
 
 internal typealias URLRequestConvertible = Alamofire.URLRequestConvertible
 
@@ -27,14 +26,14 @@ public typealias DownloadDestination = Alamofire.DownloadRequest.DownloadFileDes
 extension Request: RequestType { }
 
 /// Internal token that can be used to cancel requests
-internal final class CancellableToken: Cancellable, CustomDebugStringConvertible {
+public final class CancellableToken: Cancellable, CustomDebugStringConvertible {
     let cancelAction: () -> Void
     let request: Request?
-    fileprivate(set) var cancelled: Bool = false
+    public fileprivate(set) var cancelled: Bool = false
 
     fileprivate var lock: DispatchSemaphore = DispatchSemaphore(value: 1)
 
-    func cancel() {
+    public func cancel() {
         _ = lock.wait(timeout: DispatchTime.distantFuture)
         defer { lock.signal() }
         guard !cancelled else { return }
@@ -42,7 +41,7 @@ internal final class CancellableToken: Cancellable, CustomDebugStringConvertible
         cancelAction()
     }
 
-    init(action: @escaping () -> Void) {
+    public init(action: @escaping () -> Void) {
         self.cancelAction = action
         self.request = nil
     }
@@ -54,11 +53,33 @@ internal final class CancellableToken: Cancellable, CustomDebugStringConvertible
         }
     }
 
-    var debugDescription: String {
+    public var debugDescription: String {
         guard let request = self.request else {
             return "Empty Request"
         }
         return request.debugDescription
     }
 
+}
+
+internal typealias RequestableCompletion = (HTTPURLResponse?, URLRequest?, Data?, Swift.Error?) -> Void
+
+internal protocol Requestable {
+    func response(queue: DispatchQueue?, completionHandler: @escaping RequestableCompletion) -> Self
+}
+
+extension DataRequest: Requestable {
+    internal func response(queue: DispatchQueue?, completionHandler: @escaping RequestableCompletion) -> Self {
+        return response(queue: queue, completionHandler: { handler  in
+            completionHandler(handler.response, handler.request, handler.data, handler.error)
+        })
+    }
+}
+
+extension DownloadRequest: Requestable {
+    internal func response(queue: DispatchQueue?, completionHandler: @escaping RequestableCompletion) -> Self {
+        return response(queue: queue, completionHandler: { handler  in
+            completionHandler(handler.response, handler.request, nil, handler.error)
+        })
+    }
 }